@@ -1,62 +1,67 @@
-import 'package:HackRU/defaults.dart';
-import 'package:HackRU/models/models.dart';
-import 'package:HackRU/ui/hackru_app.dart';
-import 'package:HackRU/ui/pages/login/login_page.dart';
-import 'package:HackRU/ui/widgets/page_not_found.dart';
-import 'package:flutter/cupertino.dart';
+import 'package:hackru/defaults.dart';
+import 'package:hackru/models/models.dart';
+import 'package:hackru/ui/hackru_app.dart';
+import 'package:hackru/ui/pages/login/login_page.dart';
+import 'package:hackru/ui/widgets/page_not_found.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_local_notifications/flutter_local_notifications.dart';
-import 'package:rxdart/rxdart.dart';
+import 'package:url_strategy/url_strategy.dart';
+// import 'package:flutter_local_notifications/flutter_local_notifications.dart';
 
 import 'styles.dart';
 
-final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
-FlutterLocalNotificationsPlugin();
+// ====== TODO: update following configs
+// final FlutterLocalNotificationsPlugin flutterLocalNotificationsPlugin =
+//     FlutterLocalNotificationsPlugin();
 
-final BehaviorSubject<ReceivedNotification> didReceiveLocalNotificationSubject =
-BehaviorSubject<ReceivedNotification>();
+// final BehaviorSubject<ReceivedNotification> didReceiveLocalNotificationSubject =
+//     BehaviorSubject<ReceivedNotification>();
 
-final BehaviorSubject<String> selectNotificationSubject =
-BehaviorSubject<String>();
+// final BehaviorSubject<String> selectNotificationSubject =
+//     BehaviorSubject<String>();
 
-NotificationAppLaunchDetails notificationAppLaunchDetails;
+// NotificationAppLaunchDetails? notificationAppLaunchDetails;
 
-void main() async{
+void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
-  notificationAppLaunchDetails = await flutterLocalNotificationsPlugin
-      .getNotificationAppLaunchDetails();
+  // ====== TODO: update following configs
+  // notificationAppLaunchDetails = await flutterLocalNotificationsPlugin
+  //     .getNotificationAppLaunchDetails();
 
-  var initializationSettingsAndroid =
-  AndroidInitializationSettings('app_icon_transparent');
+  // var initializationSettingsAndroid =
+  // AndroidInitializationSettings('app_icon_transparent');
 
-  var initializationSettingsIOS = IOSInitializationSettings(
-      requestAlertPermission: false,
-      requestBadgePermission: false,
-      requestSoundPermission: false,
-      onDidReceiveLocalNotification:
-          (int id, String title, String body, String payload) async {
-        didReceiveLocalNotificationSubject.add(ReceivedNotification(
-            id: id, title: title, body: body, payload: payload));
-      });
+  // var initializationSettingsIOS = IOSInitializationSettings(
+  //     requestAlertPermission: false,
+  //     requestBadgePermission: false,
+  //     requestSoundPermission: false,
+  //     onDidReceiveLocalNotification:
+  //         (int id, String title, String body, String payload) async {
+  //       didReceiveLocalNotificationSubject.add(ReceivedNotification(
+  //           id: id, title: title, body: body, payload: payload));
+  //     });
 
-  var initializationSettings = InitializationSettings(
-      initializationSettingsAndroid, initializationSettingsIOS);
-  await flutterLocalNotificationsPlugin.initialize(initializationSettings,
-      onSelectNotification: (String payload) async {
-        if (payload != null) {
-          debugPrint('notification payload: ' + payload);
-        }
-        selectNotificationSubject.add(payload);
-      });
+  // var initializationSettings = InitializationSettings(
+  //     initializationSettingsAndroid, initializationSettingsIOS);
+  // await flutterLocalNotificationsPlugin.initialize(initializationSettings,
+  //     onSelectNotification: (String payload) async {
+  //       if (payload != null) {
+  //         debugPrint('notification payload: ' + payload);
+  //       }
+  //       selectNotificationSubject.add(payload);
+  //     });
 
-  SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
-    statusBarColor: Colors.white,
-    statusBarIconBrightness: Brightness.light,
-    systemNavigationBarColor: Colors.black,
-    systemNavigationBarIconBrightness: Brightness.light,
-  ));
+  /** ========================================================
+   *  SYSTEM UI OVERLAY STYLING (ANDROID)
+   *  ======================================================== */
+  SystemChrome.setSystemUIOverlayStyle(
+    SystemUiOverlayStyle(
+      statusBarColor: HackRUColors.black,
+    ),
+  );
+
+  setPathUrlStrategy();
   runApp(MainApp());
 }
 
@@ -64,23 +69,23 @@ void main() async{
 ///                    RECEIVED NOTIFICATION
 /// =======================================================
 
-class ReceivedNotification {
-  final int id;
-  final String title;
-  final String body;
-  final String payload;
+// class ReceivedNotification {
+//   final int? id;
+//   final String? title;
+//   final String? body;
+//   final String? payload;
 
-  ReceivedNotification({
-    @required this.id,
-    @required this.title,
-    @required this.body,
-    @required this.payload,
-  });
-}
+//   ReceivedNotification({
+//     @required this.id,
+//     @required this.title,
+//     @required this.body,
+//     @required this.payload,
+//   });
+// }
 
 class MainApp extends StatelessWidget {
-  final LcsCredential lcsCredential;
-  const MainApp({Key key, this.lcsCredential}) : super(key: key);
+  final LcsCredential? lcsCredential;
+  const MainApp({Key? key, this.lcsCredential}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
