@@ -1,57 +1,109 @@
-import 'package:HackRU/styles.dart';
-import 'package:HackRU/defaults.dart';
-import 'package:HackRU/services/hackru_service.dart';
-import 'package:HackRU/models/models.dart';
-import 'package:flare_flutter/flare_actor.dart';
+import 'package:hackru/styles.dart';
+import 'package:hackru/defaults.dart';
+import 'package:hackru/services/hackru_service.dart';
+import 'package:hackru/models/models.dart';
 import 'package:flutter/material.dart';
+
 import 'package:url_launcher/url_launcher.dart';
+
+const helpResources = [
+  {
+    "name": "MentorQ",
+    "desc": "Get matched with a mentor for help!",
+    "url": "http://mentorq.hackru.org"
+  },
+  {
+    "name": "Slack",
+    "desc": "Talk to friends, mentors, and organizers at HackRU",
+    "url":
+        "https://hackru.us3.list-manage.com/track/click?u=457c42db47ebf530a0fc733fb&id=77a8cf5266&e=c9b098417d"
+  },
+  {
+    "name": "Devpost",
+    "desc": "Submit your projects",
+    "url":
+        "https://hackru.us3.list-manage.com/track/click?u=457c42db47ebf530a0fc733fb&id=f6ef06bd11&e=c9b098417d"
+  },
+  {
+    "name": "Rutgers Campus Map",
+    "desc": "Rutgers University - College Ave Campus",
+    "url":
+        "https://maps.rutgers.edu/#/?bus=true&dining=true&healthCare=true&lat=40.503942&lng=-74.450773&parking=true&sidebar=true&zoom=17"
+  },
+  {
+    "name": "Register Your Vehicle",
+    "desc": "Register for free parking permit",
+    "url":
+        "https://hackru.us3.list-manage.com/track/click?u=457c42db47ebf530a0fc733fb&id=704cc129a9&e=c9b098417d"
+  }
+  // {
+  //   "name": "Food Menu",
+  //   "desc": "breakfast, lunch, and dinner",
+  //   "url": "https://s3-us-west-2.amazonaws.com/hackru-misc/menu.pdf"
+  // }
+];
 
 class Help extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-      body: FutureBuilder<List<HelpResource>>(
-        future: helpResources(MISC_URL),
-        builder:
-            (BuildContext context, AsyncSnapshot<List<HelpResource>> snapshot) {
-          switch (snapshot.connectionState) {
-            case ConnectionState.none:
-            case ConnectionState.waiting:
-              return Center(
-                child: Container(
-                  color: transparent,
-                  height: 400.0,
-                  width: 400.0,
-                  child: FlareActor(
-                    'assets/flare/loading_indicator.flr',
-                    alignment: Alignment.center,
-                    fit: BoxFit.contain,
-                    animation: 'idle',
-                  ),
-                ),
-              );
-            default:
-              print(snapshot.hasError);
-              var resources = snapshot.data;
-//              print(resources);
-              return Container(
-                child: ListView.builder(
-                  itemCount: resources.length,
-                  itemBuilder: (BuildContext context, int index) {
-                    return HelpButton(resource: resources[index]);
-                  },
-                ),
-              );
-          }
+      body: ListView.builder(
+        itemCount: helpResources.length,
+        itemBuilder: (BuildContext context, int index) {
+          HelpResource helpResource =
+              HelpResource.fromJson(helpResources[index]);
+          return HelpButton(
+            resource: helpResource,
+          );
         },
       ),
+//       body: FutureBuilder<List<HelpResource>>(
+//         future: helpResources(MISC_URL),
+//         builder:
+//             (BuildContext context, AsyncSnapshot<List<HelpResource>> snapshot) {
+//           switch (snapshot.connectionState) {
+//             case ConnectionState.none:
+//             case ConnectionState.waiting:
+//               return const Center(
+//                 child: CircularProgressIndicator(),
+//                 // child: Container(
+//                 //   color: HackRUColors.transparent,
+//                 //   height: 400.0,
+//                 //   width: 400.0,
+//                 //   child: const RiveAnimation.asset(
+//                 //     'assets/flare/loading_indicator.flr',
+//                 //     alignment: Alignment.center,
+//                 //     fit: BoxFit.contain,
+//                 //     animations: ['idle'],
+//                 //   ),
+//                 // ),
+//               );
+//             default:
+//               print(snapshot.hasError);
+//               var resources = snapshot.data;
+// //              print(resources);
+//               return !snapshot.hasError
+//                   ? ListView.builder(
+//                       itemCount: resources?.length,
+//                       itemBuilder: (BuildContext context, int index) {
+//                         return HelpButton(resource: resources![index]);
+//                       },
+//                     )
+//                   : const Center(
+//                       child: Text(
+//                         'An error occurred while fetching help resources!',
+//                       ),
+//                     );
+//           }
+//         },
+//       ),
     );
   }
 }
 
 class HelpButton extends StatelessWidget {
-  HelpButton({@required this.resource});
+  HelpButton({required this.resource});
   final HelpResource resource;
 
   void _open() async {
@@ -73,7 +125,7 @@ class HelpButton extends StatelessWidget {
         child: Container(
           height: 100.0,
           child: InkWell(
-            splashColor: yellow,
+            splashColor: HackRUColors.yellow,
             onTap: _open,
             child: Column(
               children: <Widget>[
@@ -85,7 +137,7 @@ class HelpButton extends StatelessWidget {
                       Text(
                         resource.name.toUpperCase(),
                         style: TextStyle(
-                          color: white,
+                          color: HackRUColors.white,
                           fontSize: 30,
                           fontWeight: FontWeight.bold,
                         ),
