@@ -1,9 +1,9 @@
-import 'package:HackRU/styles.dart';
-import 'package:HackRU/ui/pages/login/login_form.dart';
+import 'package:hackru/styles.dart';
+import 'package:hackru/ui/pages/login/login_form.dart';
 import 'package:flutter/material.dart';
 
 class LoginPage extends StatefulWidget {
-  const LoginPage({Key key}) : super(key: key);
+  const LoginPage({Key? key}) : super(key: key);
 
   @override
   _LoginPageState createState() => _LoginPageState();
@@ -20,7 +20,7 @@ class _LoginPageState extends State<LoginPage> {
     return Scaffold(
       backgroundColor: Theme.of(context).scaffoldBackgroundColor,
       appBar: AppBar(
-        backgroundColor: transparent,
+        backgroundColor: HackRUColors.transparent,
         elevation: 0.0,
         iconTheme: IconThemeData(
           color: Theme.of(context).primaryColor,
