@@ -1,33 +1,33 @@
+import 'package:hackru/styles.dart';
 import 'package:flutter/material.dart';
 import 'package:hidden_drawer_menu/hidden_drawer_menu.dart';
-import 'package:hidden_drawer_menu/menu/hidden_menu_widget.dart';
 import 'package:hidden_drawer_menu/simple_hidden_drawer/animated_drawer_content.dart';
 
 class CustomHiddenMenu extends StatefulWidget {
   /// Decocator that allows us to add backgroud in the menu(img)
-  final Widget background;
+  final Widget? background;
 
   /// that allows us to add shadow above menu items
-  final bool enableShadowItensMenu;
+  final bool? enableShadowItensMenu;
 
   /// that allows us to add backgroud in the menu(color)
-  final Color backgroundColorMenu;
+  final Color? backgroundColorMenu;
 
   /// Items of the menu
-  final List<ItemHiddenMenu> itens;
+  final List<ItemHiddenMenu>? menuItems;
 
   /// Callback to recive item selected for user
-  final Function(int) selectedListern;
+  final Function(int)? selectedListern;
 
   /// position to set initial item selected in menu
-  final int initPositionSelected;
+  final int? initPositionSelected;
 
-  final TypeOpen typeOpen;
+  final TypeOpen? typeOpen;
 
   CustomHiddenMenu(
-      {Key key,
+      {Key? key,
       this.background,
-      this.itens,
+      this.menuItems,
       this.selectedListern,
       this.initPositionSelected,
       this.backgroundColorMenu,
@@ -40,8 +40,8 @@ class CustomHiddenMenu extends StatefulWidget {
 }
 
 class _CustomHiddenMenuState extends State<CustomHiddenMenu> {
-  int _indexSelected;
-  SimpleHiddenDrawerController controller;
+  int? _indexSelected;
+  SimpleHiddenDrawerController? controller;
 
   @override
   void initState() {
@@ -51,20 +51,20 @@ class _CustomHiddenMenuState extends State<CustomHiddenMenu> {
 
   void _listenerController() {
     setState(() {
-      _indexSelected = controller.position;
+      _indexSelected = controller?.position;
     });
   }
 
   @override
   void didChangeDependencies() {
     controller = SimpleHiddenDrawerController.of(context);
-    controller.addListener(_listenerController);
+    controller?.addListener(_listenerController);
     super.didChangeDependencies();
   }
 
   @override
   void dispose() {
-    controller.removeListener(_listenerController);
+    controller?.removeListener(_listenerController);
     super.dispose();
   }
 
@@ -81,7 +81,7 @@ class _CustomHiddenMenuState extends State<CustomHiddenMenu> {
               child: Container(
                 padding: EdgeInsets.only(top: 40.0, bottom: 40.0),
                 decoration: BoxDecoration(
-                    boxShadow: widget.enableShadowItensMenu
+                    boxShadow: widget.enableShadowItensMenu!
                         ? [
                             BoxShadow(
                               color: const Color(0x44000000),
@@ -93,29 +93,46 @@ class _CustomHiddenMenuState extends State<CustomHiddenMenu> {
                         : []),
                 child: NotificationListener<OverscrollIndicatorNotification>(
                   onNotification: (scroll) {
-                    scroll.disallowGlow();
-                    return;
+                    scroll.disallowIndicator();
+                    return false;
                   },
                   child: ListView.builder(
                     shrinkWrap: true,
                     padding: EdgeInsets.all(0.0),
-                    itemCount: widget.itens.length,
+                    itemCount: widget.menuItems?.length,
                     itemBuilder: (context, index) {
-                      return HiddenMenuWidget(
-                        name: widget.itens[index].name,
+                      return ListTile(
+                        selectedColor: HackRUColors.yellow,
+                        textColor: Colors.white30,
+                        leading: Text(
+                          widget.menuItems![index].name,
+                          style: TextStyle(
+                            fontSize: 30,
+                          ),
+                        ),
                         selected: index == _indexSelected,
-                        colorLineSelected:
-                            widget.itens[index].colorLineSelected,
-                        baseStyle: widget.itens[index].baseStyle,
-                        selectedStyle: widget.itens[index].selectedStyle,
-                        typeOpen: widget.typeOpen,
                         onTap: () {
-                          if (widget.itens[index].onTap != null) {
-                            widget.itens[index].onTap();
+                          if (widget.menuItems![index].onTap != null) {
+                            widget.menuItems![index].onTap!();
                           }
-                          controller.setSelectedMenuPosition(index);
+                          controller?.setSelectedMenuPosition(index);
                         },
                       );
+                      // return HiddenMenuWidget(
+                      //   name: widget.menuItems![index].name,
+                      //   selected: index == _indexSelected,
+                      //   colorLineSelected:
+                      //       widget.menuItems![index].colorLineSelected,
+                      //   baseStyle: widget.menuItems![index].baseStyle,
+                      //   selectedStyle: widget.menuItems![index].selectedStyle,
+                      //   typeOpen: widget.typeOpen,
+                      //   onTap: () {
+                      //     if (widget.menuItems![index].onTap != null) {
+                      //       widget.menuItems![index].onTap!();
+                      //     }
+                      //     controller?.setSelectedMenuPosition(index);
+                      //   },
+                      // );
                     },
                   ),
                 ),
