@@ -3,9 +3,9 @@ import 'package:flutter/material.dart';
 import '../../../styles.dart';
 
 class ErrorDialog extends StatelessWidget {
-  final String body;
+  final String? body;
 
-  const ErrorDialog({Key key, this.body}) : super(key: key);
+  const ErrorDialog({Key? key, this.body}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -16,14 +16,14 @@ class ErrorDialog extends StatelessWidget {
       ),
       title: Icon(
         Icons.warning,
-        color: pink,
+        color: HackRUColors.pink,
         size: 80.0,
       ),
       content: Text(
-        body,
+        body!,
         style: TextStyle(
           fontSize: 18,
-          color: pink,
+          color: HackRUColors.pink,
           fontWeight: FontWeight.w700,
         ),
         textAlign: TextAlign.center,
@@ -33,7 +33,7 @@ class ErrorDialog extends StatelessWidget {
           shape: RoundedRectangleBorder(
             borderRadius: BorderRadius.circular(15.0),
           ),
-          splashColor: white,
+          splashColor: HackRUColors.white,
           height: 40.0,
           color: Theme.of(context).primaryColor,
           onPressed: () {
@@ -43,7 +43,9 @@ class ErrorDialog extends StatelessWidget {
           child: const Text(
             'OK',
             style: TextStyle(
-                fontSize: 16, color: white, fontWeight: FontWeight.w700),
+                fontSize: 16,
+                color: HackRUColors.white,
+                fontWeight: FontWeight.w700),
             textAlign: TextAlign.center,
           ),
         ),
