@@ -1,12 +1,10 @@
-import 'package:HackRU/styles.dart';
-import 'package:flare_flutter/flare_actor.dart';
-import 'package:flutter/cupertino.dart';
+import 'package:hackru/styles.dart';
 import 'package:flutter/material.dart';
 
 class PageNotFound extends StatefulWidget {
-  PageNotFound({Key key, this.title}) : super(key: key);
+  PageNotFound({Key? key, this.title}) : super(key: key);
 
-  final String title;
+  final String? title;
 
   @override
   _PageNotFoundState createState() => _PageNotFoundState();
@@ -16,7 +14,7 @@ class _PageNotFoundState extends State<PageNotFound> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: off_white,
+      backgroundColor: HackRUColors.off_white,
       body: Column(
         crossAxisAlignment: CrossAxisAlignment.center,
         mainAxisAlignment: MainAxisAlignment.center,
@@ -28,7 +26,9 @@ class _PageNotFoundState extends State<PageNotFound> {
               child: Text(
                 '404! \n Page not found!',
                 style: TextStyle(
-                    fontSize: 25.0, color: pink, fontWeight: FontWeight.w700),
+                    fontSize: 25.0,
+                    color: HackRUColors.pink,
+                    fontWeight: FontWeight.w700),
                 textAlign: TextAlign.center,
               ),
             ),
@@ -36,12 +36,13 @@ class _PageNotFoundState extends State<PageNotFound> {
           Container(
             margin: EdgeInsets.only(left: 50.0, bottom: 35.0),
             height: 400.0,
-            child: FlareActor(
-              'flare/Filip.flr',
-              alignment: Alignment.topRight,
-              fit: BoxFit.contain,
-              animation: 'idle',
-            ),
+            child: const Text('Oops, page not found!'),
+            // child: const RiveAnimation.asset(
+            //   'flare/Filip.flr',
+            //   alignment: Alignment.center,
+            //   fit: BoxFit.contain,
+            //   animations: ['idle'],
+            // ),
           ),
         ],
       ),
