@@ -6,28 +6,21 @@ part of 'models.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-LcsCredential _$LcsCredentialFromJson(Map<String, dynamic> json) {
-  return LcsCredential(
-    json['email'] as String,
-    json['token'] as String,
-    DateTime.parse(json['valid_until'] as String),
-  );
-}
+LcsCredential _$LcsCredentialFromJson(Map<String, dynamic> json) =>
+    LcsCredential(
+      json['token'] as String,
+    );
 
 Map<String, dynamic> _$LcsCredentialToJson(LcsCredential instance) =>
     <String, dynamic>{
-      'email': instance.email,
       'token': instance.token,
-      'valid_until': instance.expiration.toIso8601String(),
     };
 
-HelpResource _$HelpResourceFromJson(Map<String, dynamic> json) {
-  return HelpResource(
-    json['name'] as String,
-    json['desc'] as String,
-    json['url'] as String,
-  );
-}
+HelpResource _$HelpResourceFromJson(Map<String, dynamic> json) => HelpResource(
+      json['name'] as String,
+      json['desc'] as String,
+      json['url'] as String,
+    );
 
 Map<String, dynamic> _$HelpResourceToJson(HelpResource instance) =>
     <String, dynamic>{
@@ -36,12 +29,10 @@ Map<String, dynamic> _$HelpResourceToJson(HelpResource instance) =>
       'desc': instance.description,
     };
 
-Announcement _$AnnouncementFromJson(Map<String, dynamic> json) {
-  return Announcement(
-    text: json['text'] as String,
-    ts: json['ts'] as String,
-  );
-}
+Announcement _$AnnouncementFromJson(Map<String, dynamic> json) => Announcement(
+      text: json['text'] as String?,
+      ts: json['ts'] as String?,
+    );
 
 Map<String, dynamic> _$AnnouncementToJson(Announcement instance) =>
     <String, dynamic>{
@@ -49,30 +40,27 @@ Map<String, dynamic> _$AnnouncementToJson(Announcement instance) =>
       'ts': instance.ts,
     };
 
-User _$UserFromJson(Map<String, dynamic> json) {
-  return User(
-    json['email'] as String,
-    Role.fromJson(json['role'] as Map<String, dynamic>),
-    json['votes'] as int,
-    json['github'] as String,
-    json['major'] as String,
-    json['short_answer'] as String,
-    json['shirt_size'] as String,
-    json['first_name'] as String,
-    json['last_name'] as String,
-    json['dietary_restrictions'] as String,
-    json['special_needs'] as String,
-    json['date_of_birth'] as String,
-    json['school'] as String,
-    json['grad_year'] as String,
-    json['gender'] as String,
-    json['registration_status'] as String,
-    json['level_of_study'] as String,
-    json['dayOf'] as Map<String, dynamic>,
-  )..auth = (json['auth'] as List)
-      .map((e) => Auth.fromJson(e as Map<String, dynamic>))
-      .toList();
-}
+User _$UserFromJson(Map<String, dynamic> json) => User(
+      json['email'] as String,
+      Role.fromJson(json['role'] as Map<String, dynamic>),
+      json['votes'] as int,
+      json['github'] as String,
+      json['major'] as String,
+      json['short_answer'] as String,
+      json['shirt_size'] as String,
+      json['first_name'] as String,
+      json['last_name'] as String,
+      json['dietary_restrictions'] as String,
+      json['special_needs'] as String,
+      json['date_of_birth'] as String,
+      json['school'] as String,
+      json['grad_year'] as String,
+      json['gender'] as String,
+      json['registration_status'] as String,
+      json['level_of_study'] as String,
+      json['day_of'] as Map<String, dynamic>,
+      (json['qrcode'] as List<dynamic>?)?.map((e) => e as String).toList(),
+    );
 
 Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
       'email': instance.email,
@@ -92,21 +80,19 @@ Map<String, dynamic> _$UserToJson(User instance) => <String, dynamic>{
       'gender': instance.gender,
       'registration_status': instance.registrationStatus,
       'level_of_study': instance.levelOfStudy,
-      'dayOf': instance.dayOf,
-      'auth': instance.auth,
+      'day_of': instance.dayOf,
+      'qrcode': instance.qrcode,
     };
 
-Role _$RoleFromJson(Map<String, dynamic> json) {
-  return Role(
-    json['hacker'] as bool,
-    json['volunteer'] as bool,
-    json['judge'] as bool,
-    json['sponsor'] as bool,
-    json['mentor'] as bool,
-    json['organizer'] as bool,
-    json['director'] as bool,
-  );
-}
+Role _$RoleFromJson(Map<String, dynamic> json) => Role(
+      json['hacker'] as bool,
+      json['volunteer'] as bool,
+      json['judge'] as bool,
+      json['sponsor'] as bool,
+      json['mentor'] as bool,
+      json['organizer'] as bool,
+      json['director'] as bool,
+    );
 
 Map<String, dynamic> _$RoleToJson(Role instance) => <String, dynamic>{
       'hacker': instance.hacker,
@@ -118,14 +104,26 @@ Map<String, dynamic> _$RoleToJson(Role instance) => <String, dynamic>{
       'director': instance.director,
     };
 
-Auth _$AuthFromJson(Map<String, dynamic> json) {
-  return Auth(
-    json['token'] as String,
-    json['valid_until'] as String,
-  );
-}
+Auth _$AuthFromJson(Map<String, dynamic> json) => Auth(
+      json['token'] as String,
+      json['valid_until'] as String,
+    );
 
 Map<String, dynamic> _$AuthToJson(Auth instance) => <String, dynamic>{
       'token': instance.token,
       'valid_until': instance.validUntil,
+    };
+
+Event _$EventFromJson(Map<String, dynamic> json) => Event(
+      summary: json['summary'] as String?,
+      location: json['location'] as String?,
+      start: json['start'] == null
+          ? null
+          : DateTime.parse(json['start'] as String),
+    );
+
+Map<String, dynamic> _$EventToJson(Event instance) => <String, dynamic>{
+      'summary': instance.summary,
+      'location': instance.location,
+      'start': instance.start?.toIso8601String(),
     };
