@@ -1,8 +1,9 @@
 import 'dart:async';
 import 'dart:convert';
-import 'package:HackRU/defaults.dart';
-import 'package:HackRU/models/exceptions.dart';
-import 'package:HackRU/models/models.dart';
+import 'package:flutter/material.dart';
+import 'package:hackru/defaults.dart';
+import 'package:hackru/models/exceptions.dart';
+import 'package:hackru/models/models.dart';
 import 'package:http/http.dart' as http;
 
 var client = http.Client();
@@ -13,25 +14,29 @@ final kHeader = {'Content-Type': 'application/json'};
 ///========================================================
 
 Future<http.Response> getMisc(String endpoint) {
-  return client.get(MISC_URL + endpoint);
+  return client.get(Uri.parse(MISC_URL + endpoint));
 }
 
 Future<http.Response> getLcs(String endpoint) {
-  return client.get(BASE_URL + endpoint).timeout(const Duration(seconds: 10));
+  return client
+      .get(Uri.parse(PROD_URL + endpoint))
+      .timeout(const Duration(seconds: 10));
 }
 
 Future<http.Response> postLcs(String endpoint, dynamic kBody) async {
   var encodedBody = jsonEncode(kBody);
+  // debugPrint('try: kBody');
   var result = await client
-      .post(BASE_URL + endpoint, headers: kHeader, body: encodedBody)
+      .post(Uri.parse(BASE_URL + endpoint), headers: kHeader, body: encodedBody)
       .timeout(const Duration(seconds: 10));
 
   var decoded = jsonDecode(result.body);
+  // debugPrint('${result.statusCode} $decoded');
   if (decoded['statusCode'] != result.statusCode) {
-    print(decoded);
-    print(
-        '!!!!!!!!!!!!WARNING\nBody and Container status code dissagree actual ${result.statusCode} body: ${decoded['statusCode']}\n');
-    print(endpoint);
+    // print(decoded);
+    // print(
+    //     '!!!!!!!!!!!!WARNING\nBody and Container status code dissagree actual ${result.statusCode} body: ${decoded['statusCode']}\n');
+    // print(endpoint);
   }
   return result;
 }
@@ -64,12 +69,26 @@ Future<List<HelpResource>> helpResources(String miscUrl) async {
       .toList();
 }
 
-Future<List<String>> qrEvents(String miscUrl) async {
-  var response = await getMisc('/events.json');
-  var resources = json.decode(response.body);
-  var qrEvents = List<String>.from(resources);
-  return qrEvents;
-}
+// List<String> qrEvents(String miscUrl) async {
+// var response = await getMisc('/events.json');
+// var events = [
+//   "check-in",
+//   "check-in-no-delayed",
+//   "lunch-1",
+//   "dinner",
+//   "t-shirts",
+//   "midnight-meal",
+//   "midnight-surprise",
+//   "breakfast",
+//   "lunch-2",
+//   "raffle",
+//   "ctf-1",
+//   "ctf-2"
+// ];
+var qrEvents = json.encode(events);
+// var qrEvents = List<String>.from(resources);
+// return qrEvents;
+// }
 
 ///========================================================
 ///                     GET REQUESTS
@@ -83,6 +102,7 @@ Future<List<Announcement>> slackResources() async {
   try {
     var response = await getLcs('/dayof-slack');
     resources = json.decode(response.body);
+    // print('======== res: ' + response.body);
   } on TimeoutException catch (_) {
     return [
       Announcement(
@@ -91,11 +111,19 @@ Future<List<Announcement>> slackResources() async {
       )
     ];
   }
-  if (resources['body'].length <= 3) {
-    if (resources['body']['statusCode'] == 400) {
+  if (resources['body'].length == 0) {
+    if (resources['statusCode'] == 400) {
       return [
         Announcement(
           text: 'Error: Unable to retrieve messages!',
+          ts: tsNow,
+        )
+      ];
+    }
+    if (resources['statusCode'] == 200) {
+      return [
+        Announcement(
+          text: 'Nothing to show at the moment, please stay tuned!',
           ts: tsNow,
         )
       ];
@@ -131,6 +159,11 @@ Future<List<Event>> dayofEventsResources() async {
 
 /// TODO: Fix this
 // authorize can give wrong status codes
+// Look at LcsCredential class and figure out a way to enable commented code
+// because you'll need "email" and "token_expiry_dateTime"
+// if LCS doesn't return these info, then you'd have to call /read userData
+// right after successful /login call here and retrieve necessary userData from
+// there.
 Future<LcsCredential> login(String email, String password) async {
   var result = await postLcs('/authorize', {
     'email': email,
@@ -139,8 +172,10 @@ Future<LcsCredential> login(String email, String password) async {
   var body = jsonDecode(result.body);
   // quirk with lcs where it puts the actual result as a string
   // inside the normal body
+  // print('==== TOKEN: ${json.encode(body['body']['token'])}');
   if (body['statusCode'] == 200) {
-    return LcsCredential.fromJson(body['body']['auth']);
+    // return LcsCredential(json.encode(body['body']['token']));
+    return LcsCredential.fromJson(body['body']);
   } else if (body['statusCode'] == 403) {
     throw LcsLoginFailed();
   } else {
@@ -161,6 +196,7 @@ Future<User> getUser(String authToken, String emailAddress,
     'token': authToken,
     'query': {'email': targetEmail}
   });
+  // print('===##==== user: ${jsonDecode(result.body)}');
   if (result.statusCode == 200) {
     var users = jsonDecode(result.body)['body'];
     if (users.length < 1) {
@@ -179,14 +215,14 @@ Future<User> getUser(String authToken, String emailAddress,
 // check if the user credential belongs to is role.director first. or else it will break :(
 void updateUserDayOf(
     String lcsUrl, LcsCredential credential, User user, String event) async {
-  print(event);
+  // print(event);
   var result = await postLcs('/update', {
     'updates': {
       '\$set': {'day_of.$event': true}
     },
     'user_email': user.email,
-    'auth_email': credential.email,
-    'auth': credential.token,
+    'auth_email': user.email,
+    'token': credential.token,
   });
 
   var decoded = jsonDecode(result.body);
@@ -200,20 +236,24 @@ void updateUserDayOf(
   }
 }
 
-Future<int> attendEvent(String lcsUrl, LcsCredential credential,
+Future<int> attendEvent(String lcsUrl, String _authE, String _authT,
     String userEmailOrId, String event, bool again) async {
-  print(event);
+  // debugPrint(event);
+  // debugPrint(
+  //     'Attend even called with \n auth: $_authE \n token: $_authT \n for user: $userEmailOrId \n for the event: $event');
   var result = await postLcs('/attend-event', {
-    'auth_email': credential.email,
+    'auth_email': _authE, //TODO: figure out how to get auth_email here
     'qr': userEmailOrId,
-    'token': credential.token,
+    'token': _authT,
     'event': event,
     'again': again
   });
-
-  var body = jsonDecode(result.body);
+  // debugPrint('Attend event received ${result.body}');
+  var body = json.decode(result.body);
   if (body['statusCode'] == 200) {
-    return body['body']['_count'];
+    var resp = body['body']['new_count'];
+    // debugPrint('====== in 200 responsebody $resp');
+    return resp;
   } else if (body['statusCode'] == 402) {
     throw UserCheckedEvent();
   } else if (body['statusCode'] == 404) {
@@ -223,13 +263,14 @@ Future<int> attendEvent(String lcsUrl, LcsCredential credential,
   }
 }
 
-void linkQR(String lcsUrl, LcsCredential credential, String userEmailOrId,
+void linkQR(String lcsUrl, String _authE, String _authT, String userEmailOrId,
     String hashQR) async {
-  print(hashQR);
+  // debugPrint('linkQR called: $hashQR');
+  // debugPrint('$_authE $_authT $userEmailOrId');
   var result = await postLcs('/link-qr', {
-    'auth_email': credential.email,
+    'auth_email': _authE, // figure out how to get auth email here
     'email': userEmailOrId,
-    'token': credential.token,
+    'token': _authT,
     'qr_code': hashQR
   });
 
@@ -241,4 +282,26 @@ void linkQR(String lcsUrl, LcsCredential credential, String userEmailOrId,
   } else if (decoded['statusCode'] != 200) {
     throw LcsError(result);
   }
+}
+
+Future<bool> isAuthorizedForQRScanner(String token, String email) async {
+  var userProfile = await getUser(token, email);
+  if (userProfile.role.organizer == true ||
+      userProfile.role.organizer == true) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+void main() {
+  Future<void> testCall(email, password) async {
+    LcsCredential cred = await login(email, password);
+    // print('====== token/email: ${cred.token} / $email');
+
+    User user = await getUser(cred.token, email);
+    // print('====== user: ${user.toJson()}');
+  }
+
+  testCall('a@a.com', 'a');
 }
